# Initial hardware-accellerated aarch64 support