--- a/benches/rand.rs
+++ b/benches/rand.rs
@@ -1,11 +1,11 @@
 #[macro_use]
 extern crate criterion;
-extern crate rand;
 extern crate crc32c;
+extern crate rand;
 
-use criterion::{Criterion, Benchmark, Throughput};
-use rand::{rngs::OsRng, RngCore};
 use crc32c::crc32c;
+use criterion::{Benchmark, Criterion, Throughput};
+use rand::{rngs::OsRng, RngCore};
 
 fn crc32c_megabyte(c: &mut Criterion) {
     let mut bytes = [0u8; 1_000_000];
@@ -13,9 +13,8 @@ fn crc32c_megabyte(c: &mut Criterion) {
 
     c.bench(
         "crc32_update_megabytes",
-        Benchmark::new("crc32_update_megabytes", move |b| {
-            b.iter(|| crc32c(&bytes))
-        }).throughput(Throughput::Bytes(1_000_000)),
+        Benchmark::new("crc32_update_megabytes", move |b| b.iter(|| crc32c(&bytes)))
+            .throughput(Throughput::Bytes(1_000_000)),
     );
 }