--- a/build.rs
+++ b/build.rs
@@ -1,9 +1,8 @@
 extern crate rustc_version;
 
-use std::{mem, ops, io};
-use std::path::Path;
 use rustc_version::{version_meta, Channel};
-
+use std::path::Path;
+use std::{io, mem, ops};
 
 /// CRC-32-Castagnoli polynomial in reversed bit order.
 pub const POLYNOMIAL: u32 = 0x82_F6_3B_78;
@@ -146,6 +145,7 @@ fn hw_table(len: usize) -> [[u32; 256]; 4] {
     zeroes
 }
 
+// LONG/SHORT VALUES MUST BE SYNCHRONIZED WITH src/tables.rs
 pub const LONG: usize = 8192;
 pub const SHORT: usize = 256;