Somewhat naive implementation for using the crc32c instructions on arm64 so far, but the remaining parts of this are:
- [x] Remove software implementation for the outer bytes
- [x] Parallelize crc32cd computations

**BENCHMARKS**
Performed on a late 2020 M1 Macbook Air
 
Before:
```
crc32_update_megabytes  time:   [664.02 us 664.58 us 665.24 us]
                        thrpt:  [1.4000 GiB/s 1.4014 GiB/s 1.4026 GiB/s]
```

After:
```
crc32_update_megabytes  time:   [92.517 us 92.865 us 93.245 us]
                        thrpt:  [9.9879 GiB/s 10.029 GiB/s 10.067 GiB/s]
                 change:
                        time:   [-86.051% -85.998% -85.928%] (p = 0.00 < 0.05)
                        thrpt:  [+610.65% +614.19% +616.91%]
```